Remove apple-mobile-web-app-capable tag to work around Safari session issues
iOS Safari does not send session cookies when following links to PWAs using
the apple-mobile-web-app-capable tag. This is not documented on
<https://developer.apple.com/library/archive/documentation/AppleApplications/Reference/SafariWebContent/ConfiguringWebApplications/ConfiguringWebApplications.html>,
though I can't say whether it is a bug in Safari or lacking documentation.

See <https://github.com/foosel/OctoPrint/issues/1792> for a more detailed
description of this type of issue.