Simplify development via docker compose
YAML anchors are nice, but also a bit hard to read. Instead of that,
we're using a dedicated .env file with several values. Also, we provide
a simple no-op configuration that's used as the base.

That sample configuration reads the necessary configuration from the
passed environment variables and therefore makes it easy to adapt for
other container-based deployments, while offering an entry point for
contributions.