dbris: show checkin suggestions at the top of the departure board
Suggestions for connecting trains (i.e., suggestions shown in the "checked in"
view) are not supported yet. This part is due for a major rewrite that moves
the associated departure board request to the background worker rather than
firing off a promise and delaying rendering until it has been answered.